% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{botregions}
\alias{botregions}
\title{Countries and associated classification of botanical divisions}
\format{
A \code{data.frame} with the following columns:
\describe{
\item{country}{Character. Botanical Country names (Level 3)}
\item{botanical_division}{Character. Basic Recording Unit names (Level 4)}
\item{continent}{Character. Continental level names (Level 1)}
\item{country_EN, country_ES, country_FR, country_PT}{Character. Botanical
Country names in English, Spanish, French, and Portuguese}
\item{botanical_division_EN, botanical_division_ES, botanical_division_FR,
botanical_division_PT}{Character. Basic Recording Unit names in
English, Spanish, French, and Portuguese}
\item{continent_EN, continent_ES, continent_FR, continent_PT}{Character.
Continental names in English, Spanish, French, and Portuguese}
}
}
\source{
\url{https://www.tdwg.org/standards/wgsrpd/}

Brummitt, R.K. (2001). World Geographic Scheme for Recording Plant Distributions,
Edition 2. Hunt Institute for Botanical Documentation, Carnegie Mellon University,
Pittsburgh.
}
\usage{
data(botregions)
}
\description{
Countries and associated classification of botanical divisions according to
the World Geographical Scheme for Recording Plant Distributions (WGSRPD).
This dataset provides the tabular look-up table with multilingual names for
all botanical regions recognized by the standard.
}
\examples{
\dontrun{
library(aRboretum)
data(botregions)
head(botregions)

# Show botanical divisions in South America (Portuguese names)
subset(botregions, continent_EN == "South America",
       select = c(botanical_division_PT, country_PT))

# Count botanical countries by continent
table(botregions$continent_EN)
}

}
