% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sigmoid.R
\name{geom_sigmoid}
\alias{geom_sigmoid}
\title{Geom sigmoid}
\usage{
geom_sigmoid(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  n = 20,
  c1 = 0.1,
  c2 = 0,
  inherit.aes = TRUE,
  ...
)
}
\description{
Geom sigmoid
}
\examples{
library(tidyverse)
df = tribble(
    ~x1, ~x2, ~y1, ~y2, ~col,
    0, 1, 0, 1, 'blue',
    0, 1, 0, -1, 'green',
)
df \%>\%
ggplot() +
geom_sigmoid(aes(x=x1, xend=x2, y=y1, yend=y2, col=col), c1=0.001, c2=.5, n=101) +
scale_colour_manual(values = df$col \%>\% set_names(df$col)) +
geom_point(aes(x=x1, y=y1)) +
geom_point(aes(x=x2, y=y2), col='red')
}
