% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{load_tree_data}
\alias{load_tree_data}
\title{Load tree data}
\usage{
load_tree_data()
}
\value{
A tibble
}
\description{
Load tree data
}
\examples{
df = load_tree_data()
df \%>\% get_roots()

taxon = 'theropoda'

df \%>\% is_root(taxon)
df \%>\% get_parent(taxon)
df \%>\% get_siblings(taxon)
df \%>\% get_children(taxon)

df2 = df \%>\% subset_taxon(taxon)
df2 \%>\% is_root(taxon)
df2 \%>\% get_parent(taxon)
df2 \%>\% get_siblings(taxon)
df2 \%>\% get_children(taxon)

df3 = df2 \%>\% collapse_taxa('tetanurae')
df3 \%>\% print(n=Inf)
df3 \%>\% get_tips()
df3 \%>\% get_nodes()
df3
df3 \%>\% arrange_tree()

}
