% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data.R
\name{subset_taxon}
\alias{subset_taxon}
\title{Subset taxon}
\usage{
subset_taxon(df, taxon)
}
\arguments{
\item{df}{a data frame of tree data (tibble)}

\item{taxon}{taxonomic name (string)}
}
\value{
A tibble
}
\description{
Get a subset of tree data for a particular taxon
}
