% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{plot tree data}
\usage{
tree(taxon = NULL, collapse = NULL, ...)
}
\arguments{
\item{taxon}{TODO}

\item{collapse}{TODO}

\item{...}{TODO}

\item{max_tips}{TODO}
}
\description{
plot tree data
}
\examples{
library(arboretum)
tree()
tree('dinosauria')
tree('dinosauria',
	 collapse = c('sauropodomorpha',
	 			 'ornithopoda',
	 			 'theropoda',
	 			 'ankylosauria',
	 			 'stegosauria',
	 			 'pachycephalosauria',
	 			 'ceratopsia'))
tree('ornithischia', collapse='none')
tree('sauropoda', collapse='none')
tree('tetanurae', collapse='avialae')
tree('sauropterygia', collapse='none')
tree('synapsida')
}
