% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_old.R
\name{tree_old}
\alias{tree_old}
\title{plot tree data}
\usage{
tree_old(taxon = NULL, collapse = NULL, max_tips = 300, ...)
}
\arguments{
\item{taxon}{TODO}

\item{collapse}{TODO}

\item{max_tips}{TODO}

\item{...}{TODO}
}
\description{
plot tree data
}
\examples{
library(arboretum)
tree()
tree('dinosauria')
tree('dinosauria',
	 collapse = c('sauropodomorpha',
	 			 'ornithopoda',
	 			 'theropoda',
	 			 'ankylosauria',
	 			 'stegosauria',
	 			 'pachycephalosauria',
	 			 'ceratopsia'))
tree('ornithischia', collapse='none')
tree('sauropoda', collapse='none')
tree('tetanurae', collapse='avialae')
tree('sauropterygia', collapse='none')
tree('synapsida')
}
