% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{plot tree data}
\usage{
tree(taxon = NULL, collapse = NULL, max_tips = 100)
}
\arguments{
\item{taxon}{TODO}

\item{collapse}{TODO}

\item{max_tips}{TODO}
}
\description{
plot tree data
}
\examples{
library(arboretum)
tree()
tree('amniota', collapse=c('archosauromorpha',
                           'therapsida',
                           'lepidosauromorpha'))
tree('therapsida', collapse='mammalia')
tree('mammalia')
tree('archosauromorpha', collapse=c('dinosauria',
                                    'pterosauria',
                                    'pseudosuchia',
                                    'sauropterygia'))
tree('pseudosuchia')
tree('dinosauria', collapse = c('sauropoda',
                                'ornithopoda',
                                'theropoda',
                                'ankylosauria',
                                'ceratopsia'))
tree('theropoda', collapse=c('avialae'))
}
