% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{plot tree data}
\usage{
tree(taxon = NULL, collapse = NULL, ...)
}
\arguments{
\item{taxon}{TODO}

\item{collapse}{TODO}

\item{...}{TODO}

\item{max_tips}{TODO}
}
\description{
plot tree data
}
\examples{
library(arboretum)
tree('tetrapodomorpha')
tree('reptiliomorpha', collapse=c('diapsida', 'therapsida'), xmin=-345, xmax=-252)
tree('therapsida', xmax=-200)
tree('ornithischia', collapse='none')
tree('sauropoda', collapse='none')
tree('tyrannosauroidea', collapse='avialae', xmin=-173)
}
